[469299][469302][469305] File viewer delete button fix, CSV Editor loading issue, and mouse wheel zooming for VisIt.
The file viewer button now removes from ResourceComponent instead of a
local arraylist. The CSV Editor's path has been repaired so that it can
be properly loaded. The VisIt editors now receive focus properly, and
can use the mouse wheel for zooming. Added error messages when user
attempts to open a VisIt connection on a port that is already in use. 

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=469299
Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=469302
Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=469305
Signed-off-by: Robert Smith <r8s@ornl.gov>