Adding to Feature/Product, improving Item error handling
Added remote jsch ui bundle to the feature and product so that we can
add a new remote connection for remote job launching. Added a protected
method to Item for better error handling. Subclasses can now invoke this
method to throw up a descriptive error dialog. Set up MOOSELauncher to
check that the required files are in the project space, and if not throw
up an error. 

Also fixed bug with Core and XMLPP, lowered XMLPP bundle start level to
4 instead of 5. 

Signed-off-by: Alex McCaskey <mccaskeyaj@ornl.gov>