Fixed some problems in the Product and Feature.
I added Mylyn bundles and all required bundles to the Product. I also
cleaned up a few more bundle versions to prevent the target from failing
in the build. I removed egit and jgit from the feature and put them into
the product as features. There is a lot more work that needs to be done
to clean up both of these files.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>