Fixing bug 481431 to open proper ICEFormEditor for ICE XML Items
This commit modifies the XMLFormContentDescriber to ensure that the
correct ICEFormEditor subclass is used when a users double-clicks the
ICE Item XML file. It checks that no other subclass of
XMLFormContentDescriber is appropriate for describing the incoming text.
If not, it returns VALID and the ICEFormEditor is used, otherwise, the
correct ICEFormEditor subclass corresponding to the VALID
XMLFormContentDescriber is used. 

Added a ReflectivityXMLFormContentDescriber and tester. Still need one
for BatMLFormEditor and ReactorFormEditor

Signed-off-by: amccaskey <mccaskeyaj@ornl.gov>