Implemented JavaFX Mesh Editor
Created a IVizMeshCanvas interface for use by the UI when interacting
with Mesh Editor canvases. The JME3 Mesh Editor now implements this
interface. 

The JavaFX Mesh Editor is now working, although there needs to be a
refactor for the observer pattern to recognize different kinds of
messages, to both fix a bug with the Mesh Element treee viewer and for
performance issues.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>