Updated tests and fixed some Mesh Editor bugs
Updated all tests to function with the newly refactored data types.

Also fixed two Mesh Editor bugs. One, caused by drawing from the generic
FXViewer, was causing its composite not to be drawn to full size. The
second allowed a user to keep adding vertices to a polygon indefinitely
if it was selected through the Mesh Elements view. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>