Triangle refactor and added FXViewer context menu
Changed the Triangle class so that it maintains separate Vertex1,
Vertex2, and Vertex3 fields rather than an EList of Vertices. This was
done so that Triangles in a mesh may be edited, as it is impossible for
one EObject to be contained within two different ELists. 

Also added a context menu to the FXViewer. It is currently populated
with an option to reset the camera to the default position. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>