Improved ColorDecorator
Move the JavaFX classes from ColorDecorator to FXColorDecorator. Also
changed the decorator's priorities so that it will display a set RGB
color over a set material. This is done so that the decorator can have
its color temporarily changed without having to set the material to null
and then replace it later.

Signed-off-by: Robert Smith <smithrw@ornl.gov>