Updated and cleaned up geometry.* plugins
Updated the geometry.* plugins to the last version from EAVP. This was
mainly changing from OSGI services to extension points. 

Cleaned up the code in response to bug #26. This included adding missing
xtexd-gen folders to the repo and removing the ui and editor code,
leaving only the data structures themselves. 

Signed-off-by: Robert Smith <smithrw@ornl.gov>